+++ cmd/vroom/config.go
@@ -10,6 +10,9 @@ type (
 		OccurrencesKafkaBrokers []string `env:"SENTRY_KAFKA_BROKERS_OCCURRENCES" env-default:"localhost:9092"`
 		OccurrencesKafkaTopic   string   `env:"SENTRY_KAFKA_TOPIC_OCCURRENCES"   env-default:"ingest-occurrences"`
 
+		SpansKafkaBrokers        []string `env:"SENTRY_KAFKA_BROKERS_SPANS" env-default:"localhost:9092"`
+		MetricsSummaryKafkaTopic string   `env:"SENTRY_KAFKA_TOPIC_METRICS_SUMMARY"   env-default:"snuba-metrics-summaries"`
+
 		ProfilingKafkaBrokers []string `env:"SENTRY_KAFKA_BROKERS_PROFILING" env-default:"localhost:9092"`
 		CallTreesKafkaTopic   string   `env:"SENTRY_KAFKA_TOPIC_CALL_TREES"  env-default:"profiles-call-tree"`
 		ProfilesKafkaTopic    string   `env:"SENTRY_KAKFA_TOPIC_PROFILES"    env-default:"processed-profiles"`